+++ contracts/interfaces/IVotingPower.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.7.0;
 
 interface IVotingPower {
     function stake(uint256 amount) external;
-    function stakeWithPermit(uint256 amount, uint deadline, uint8 v, bytes32 r, bytes32 s) external;
+    function stakeWithPermit(uint256 amount, uint256 deadline, uint8 v, bytes32 r, bytes32 s) external;
     function withdraw(uint256 amount) external;
     function addVotingPowerForVestingTokens(address account, uint256 amount) external;
     function removeVotingPowerForClaimedTokens(address account, uint256 amount) external;
